% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfabric.R
\name{bfabric}
\alias{bfabric}
\title{shiny server module for the bfabric}
\usage{
bfabric(input, output, session, applicationid)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{
check the \code{input$resourceid} value.
}
\description{
provides a shiny server module for the bfabric system.
It is assumes that the \code{exec/flask_bfabric_sample.py} programm is running.

\code{ssh-keygen -f $PWD/bfabricShiny.key -t rsa} will generate 
the key files.
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2017
}
\references{
\url{https://doi.org/10.1145/1739041.1739135}
}
\seealso{
\url{http://fgcz-bfabric.uzh.ch}
}

