% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{bfabric_save}
\alias{bfabric_save}
\title{Title}
\usage{
bfabric_save(login, webservicepassword, endpoint = "workunit", query)
}
\arguments{
\item{login}{}

\item{webservicepassword}{}

\item{endpoint}{}

\item{query}{}
}
\description{
Title
}
