% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{bfabric_upload_file}
\alias{bfabric_upload_file}
\title{upload resouce to bfabric}
\usage{
bfabric_upload_file(
  login,
  webservicepassword,
  projectid = 1000,
  applicationid = 217,
  status,
  description = "",
  inputresource = NULL,
  workunitname = "MaxQuant result",
  resourcename = "MaxQuant report",
  file_content = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword}

\item{projectid}{a project or order id}

\item{applicationid}{the application id}

\item{status}{'pending', 'failed', or 'available'}

\item{workunitname}{the workunit name}

\item{resourcename}{the reosurce name}

\item{file_content}{a BLOB containing the content}
}
\value{

}
\description{
upload resouce to bfabric
}
