% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfabric.R
\name{bfabricInput}
\alias{bfabricInput}
\title{defines the bfabric shiny UI module}
\usage{
bfabricInput(id)
}
\arguments{
\item{id}{shiny session id}
}
\value{
tagList
}
\description{
defines the bfabric shiny UI module
}
\examples{
\dontrun{

    mainPanel(
      tabsetPanel(
        tabPanel("bfabric", bfabricInput("bfabric8")),
        tabPanel("plot", plotOutput("distPlot"))
   )

}

}
\references{
\url{https://doi.org/10.1145/1739041.1739135}
}
\seealso{
\url{http://fgcz-bfabric.uzh.ch}
}
