% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createWorkunit}
\alias{createWorkunit}
\title{Generate a workunit and upload a list of resource(s) (file)
to the internal bfabric storage}
\usage{
createWorkunit(
  login = NULL,
  webservicepassword = NULL,
  posturl = "http://127.0.0.1:5000",
  containerid = 3000,
  applicationid = 212,
  workunitname = "bfabricShiny result",
  parameters = structure(list(), names = character(0)),
  files = structure(list(), names = character(0)),
  inputresourceid = list(),
  links = structure(list(), names = character(0)),
  description = "",
  posturlsuffix = "create/workunit/v1"
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}

\item{containerid}{a containerid (project id or order id)}

\item{applicationid}{a application id}

\item{workunitname}{the workunit name}

\item{inputresourceid}{an integer or a list of integer of inputresourceIds.
Of note, this works only for  succeeding/preceding applications.
Default is set to \code{NULL}.}

\item{description}{free text, default is is empty.}

\item{posturlsuffix}{defines the method to use, e.g.,
read. also, save should work}

\item{status}{in \code{c('AVAILABLE', 'FAILED', 'PENDING')}
default is 'PENDING'.}

\item{resourcename}{the reosurce name}

\item{file}{a filename for a file to be uploaded.}
}
\value{
returns a nested list containing the workunit

object returned by the save method.
}
\description{
Generate a workunit and upload a list of resource(s) (file)
to the internal bfabric storage
}
\examples{

## create a csv file
tf <- tempfile(fileext = ".csv", pattern = "R-iris-data-")
write.csv(iris, file = tf)


createWorkunit(
  login = login,
  webservicepassword = webservicepassword,
  posturl = bfabricposturl,
  containerid = 3000,
  applicationid = 212,
  workunitname = "TEST Xerces",
  files = list(tf),
  description = "ignore",
  posturlsuffix = "/create/workunit/v1"
) -> rv

print(rv)
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2016-2023, MdE 2023-03-17, LS, 2025-12-15
}
