% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{.generate_folder_name}
\alias{.generate_folder_name}
\title{Title}
\usage{
.generate_folder_name(x, foldername, area = "Proteomics",
  instrument = "FUSION_1", username = "bfabricusername",
  pathprefixsep = "/")
}
\arguments{
\item{x}{}

\item{foldername}{}

\item{area}{}

\item{instrument}{}

\item{username}{}

\item{pathprefixsep}{}
}
\description{
Title
}
