% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{.generate_template_base}
\alias{.generate_template_base}
\title{.generate_template_base}
\usage{
.generate_template_base(x)
}
\arguments{
\item{x:}{the sample information (data.frame)}
}
\value{
the original dataframe (data.frame)
}
\description{
.generate_template_base
}
