% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.getApplications}
\alias{.getApplications}
\title{getApplications}
\usage{
.getApplications(login, webservicepassword, posturl)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}
}
\value{
list of bfabric applications
}
\description{
getApplications
}
\examples{
\dontrun{
library(bfabricShiny)
A <- .getApplications(login, webservicepassword)
bfabricApplication <- data.frame(id = sapply(A, function(x){x$`_id`}),
 name = sapply(A, function(x){x$name}))
bfabricApplication <- bfabricApplication[order(bfabricApplication$id),]
write.table(bfabricApplication, file="./inst/extdata/application.csv",
  row.names = FALSE)
}

}
