% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.getResources}
\alias{.getResources}
\title{get all resources of a (login, project)}
\usage{
.getResources(
  login = NULL,
  webservicepassword = NULL,
  posturl = NULL,
  workunitid = NULL,
  updateProgress = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}

\item{workunitid}{a workunit Id}

\item{updateProgress}{a callback function for writing log output, e.g.,
using a \code{\link[shiny]{Progress}} object,
see also \url{https://shiny.rstudio.com/articles/progress.html}.}
}
\value{
a vector of resource ids
}
\description{
get all resources of a (login, project)
}
