% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.getWorkunits}
\alias{.getWorkunits}
\title{CP 2024-12-24}
\usage{
.getWorkunits(
  login = NULL,
  webservicepassword = NULL,
  posturl = NULL,
  containerid = 3000,
  applicationid = 224,
  updateProgress = NULL,
  createdbefore = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}

\item{updateProgress}{a callback function for writing log output, e.g.,
using a \code{\link[shiny]{Progress}} object,
see also \url{https://shiny.rstudio.com/articles/progress.html}.}
}
\description{
CP 2024-12-24
}
