% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.listToXml}
\alias{.listToXml}
\title{List to XML}
\usage{
.listToXml(item, tag)
}
\description{
Convert List to XML
}
\details{
Can convert list or other object to an xml object using xmlNode
}
\author{
David LeBauer, Carl Davidson, Rob Kooper
}
