% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.read}
\alias{.read}
\title{.read}
\usage{
.read(
  login = NULL,
  webservicepassword = NULL,
  endpoint = "workunit",
  page = 1,
  query = list(),
  posturl = NULL,
  posturlsuffix = "read",
  idonly = FALSE,
  updateProgress = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{page}{define requested page, default is 1}

\item{query}{e,g, \code{list(containerid = 3000)}}

\item{posturl}{where the flask server is working}

\item{posturlsuffix}{defines the method to use, e.g., read. also, save should work}

\item{updateProgress}{a callback function for writing log output, e.g.,
using a \code{\link[shiny]{Progress}} object,
see also \url{https://shiny.rstudio.com/articles/progress.html}.}
}
\description{
.read
}
\author{
MdE/CP 2023-03
}
