% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.readCustomAttributes}
\alias{.readCustomAttributes}
\title{reads Custom Attributes of a workunit}
\usage{
.readCustomAttributes(login, webservicepassword, workunitid = 202570)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{workunitid}{a valid bfabric workunit id}
}
\value{
a \code{data.frame} containing columns sampleId , resourceId, and
iff exisiting a merged table of the Custom Attributes.
}
\description{
reads Custom Attributes of a workunit
}
