% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{.tray_position}
\alias{.tray_position}
\title{Title}
\usage{
.tray_position(
  queue,
  startpos = list(row = 1, col = "A", plate = 1),
  instrument = ""
)
}
\arguments{
\item{instrument}{}
}
\value{

}
\description{
Title
}
\examples{
bfabricShiny::.tray_position(bfabricShiny:::.test_data_medium(), instrument="LUMOS_2")

}
