% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.validateToken}
\alias{.validateToken}
\title{Validate Token}
\usage{
.validateToken(token, posturl = .posturl())
}
\arguments{
\item{token}{a given token}

\item{posturl}{the bfabric REST proxy URL}
}
\value{
login webservicepassword
}
\description{
Validate Token
}
\author{
LS,CP 2025-07-03 devDay
}
