% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{.waters}
\alias{.waters}
\title{method eksigent}
\usage{
.waters()
}
\description{
method eksigent
}
\examples{
r <- bfabricShiny:::.waters()
r
all.equal(r[[2]][1:5] , c("\"1:A,1\"", "\"1:A,2\"", "\"1:A,3\"", "\"1:A,4\"", "\"1:A,5\""))

}
