% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{generate_queue}
\alias{generate_queue}
\title{MS instrument queue generation}
\usage{
generate_queue(x, foldername = "", projectid = 1000, area = "Proteomics",
  instrument = "FUSION_1", username = "cpanse", how.often = 2,
  how.many = 1, multiple = 1, hplc = list(name = "easylc", clean = "F6",
  standard = "F8"), method = "default", pathprefix = "D:\\\\Data2San",
  pathprefixsep = "\\\\")
}
\arguments{
\item{x}{}

\item{foldername}{}

\item{projectid}{}

\item{area}{}

\item{instrument}{}

\item{username}{}

\item{how.often}{}

\item{how.many}{}

\item{multiple}{}

\item{hplc}{}

\item{method}{}

\item{pathprefix}{}

\item{pathprefixsep}{}
}
\value{
a table which contains a MS instrument queue configuration.
sample.id' = res$extract.sampleid
}
\description{
MS instrument queue generation
}
\examples{

generate_queue(bfabricShiny:::test_data())

S <- as.data.frame(fromJSON("http://localhost:5000/extract/1000"))
generate_queue(S)

}

