% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{getHPLCparameter}
\alias{getHPLCparameter}
\title{Define HPLC, autoQC01, autoQC02, autoQC4L, clean ----
list elements are: 1) HPLC, 2) sample positions, 3) autoQC01 position, 4) autoQC02 position, 5) autoQC4L position, 6) clean position}
\usage{
getHPLCparameter(row = 1, col = "A", plate = 1)
}
\description{
Define HPLC, autoQC01, autoQC02, autoQC4L, clean ----
list elements are: 1) HPLC, 2) sample positions, 3) autoQC01 position, 4) autoQC02 position, 5) autoQC4L position, 6) clean position
}
\examples{

bfabricShiny:::getHPLCparameter()[["LUMOS_2"]]
bfabricShiny:::getHPLCparameter(3,"D",2)[["LUMOS_2"]]
}
