% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{getProjects}
\alias{getProjects}
\title{queries projects of a login}
\usage{
getProjects(login, webservicepassword)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword}
}
\value{
a vector of project ids
}
\description{
queries projects of a login
}
