% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{getResources}
\alias{getResources}
\title{get all resources of a (login, project)}
\usage{
getResources(login = NULL, webservicepassword = NULL, workunitid = NULL)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword}

\item{project}{a project or order Id}
}
\value{
a vector of resource ids
}
\description{
get all resources of a (login, project)
}
