% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read}
\alias{read}
\title{read function which supports pages}
\usage{
read(
  login = NULL,
  webservicepassword = NULL,
  endpoint = "workunit",
  offset = 0,
  maxitems = 100,
  query = list(),
  posturl = NULL,
  posturlsuffix = "read",
  idonly = FALSE,
  updateProgress = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{maxitems}{max number of supported pages to}

\item{query}{e,g, \code{list(containerid = 3000)}}

\item{posturl}{where the flask server is working}

\item{posturlsuffix}{defines the method to use, e.g.,
read. also, save should work}

\item{updateProgress}{a callback function for writing log output, e.g.,
using a \code{\link[shiny]{Progress}} object,
see also \url{https://shiny.rstudio.com/articles/progress.html}.}

\item{page}{define requested page, default is 1}
}
\description{
read function which supports pages
}
\examples{
\dontrun{
bfabricShiny::read(login,
  webservicepassword,
  endpoint = 'sample',
  query = list('containerid' = 34777),
  posturl = bfabricposturl,
  maxitems = 1000) -> rv
  }
}
\author{
MdE/CP 2023-03; CP 2024-12-24; LS 2025-12-15
}
