% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{readCustomAttributes}
\alias{readCustomAttributes}
\title{reads Custom Attributes of a}
\usage{
readCustomAttributes(login, webservicepassword, workunitid = 202570)
}
\arguments{
\item{login}{}

\item{webservicepassword}{}

\item{workunitid}{}
}
\value{
a \code{data.frame} containing columns sampleId , resourceId, and
iff exisiting a merged table of the Custom Attributes.
}
\description{
reads Custom Attributes of a
}
