% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{readPages}
\alias{readPages}
\title{read function which supports pages}
\usage{
readPages(
  login = NULL,
  webservicepassword = NULL,
  endpoint = "workunit",
  offset = 0,
  maxitems = 100,
  query = list(),
  posturl = NULL,
  posturlsuffix = "read",
  idonly = FALSE,
  updateProgress = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{query}{e,g, \code{list(containerid = 3000)}}

\item{posturl}{where the flask server is working}

\item{updateProgress}{a callback function for writing log output, e.g.,
using a \code{\link[shiny]{Progress}} object,
see also \url{https://shiny.rstudio.com/articles/progress.html}.}

\item{maxpages}{max number of supported pages to}
}
\value{
a list
}
\description{
read function which supports pages
}
\examples{
# ensure you have login, webservicepassword, and posturl defined
Rprofile <- file.path(Sys.getenv("HOME"), ".Rprofile")
source(Rprofile, local = TRUE)

bfabricShiny::readPages(login, webservicepassword , endpoint = 'user',
query=list(login='cpanse'), posturl = bfabricposturl)

bfabricShiny::readPages(login,
  webservicepassword,
  endpoint = 'user',
  query=list(),
  posturl = bfabricposturl,
  updateProgress =  function(...){cat(...)})

bfabricShiny::readPages(login, webservicepassword , endpoint = 'user',
query=list(login='cpanse'), posturl = bfabricposturl)

\dontrun{
  fraction <- bfabricShiny::readPages(login = login,
    webservicepassword = webservicepassword,
    endpoint = 'sample',
    query = list( attribute = list(name = 'fraction',
        value = 'true')))
        }
}
