% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{readPlate}
\alias{readPlate}
\title{Read plate}
\usage{
readPlate(plateid, login, webservicepassword, posturl)
}
\arguments{
\item{plateid}{plate id}

\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}
}
\description{
extracts plate and sample information from the plate id
returns a data.frame containing: sample ID, sample name, tube ID, position
}
\examples{
readPlate(plateid = 2616, login, webservicepassword, bfabricposturl)
}
