% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{readSample}
\alias{readSample}
\title{Read sample
returns data.frame containing: sample name, filename, sample ID}
\usage{
readSample(sampleIds, login, webservicepassword, posturl, maxitems = 2000)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}

\item{samplelist}{list of sample ids}
}
\description{
Read sample
returns data.frame containing: sample name, filename, sample ID
}
