% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{runQueue}
\alias{runQueue}
\title{Run shiny queue generator application}
\usage{
runQueue(ipadress = "172.23.255.1")
}
\value{

}
\description{
Run shiny queue generator application
}
\examples{
#bfabricShiny::runQueue()
}
