% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{save}
\alias{save}
\title{Saves an object in bfabric}
\usage{
save(login, webservicepassword, endpoint = "workunit", query)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{query}{e,g, \code{list(containerid = 3000)}}
}
\value{
bfabric json object.
}
\description{
Saves an object in bfabric
}
\examples{
\dontrun{
bfabricShiny::save(login, webservicepassword , 'workunit',
  list(id=254893, description='test2',
  inputresourceid=list(1753925, 1753924))
}
}
