% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfabric.R
\name{bfabric}
\alias{bfabric}
\title{shiny server module for the bfabric}
\usage{
bfabric(
  input,
  output,
  session,
  applicationid = NULL,
  resoucepattern = ".*",
  resourcemultiple = FALSE
)
}
\arguments{
\item{input}{default module input}

\item{output}{default module output}

\item{session}{module session}
}
\value{
check the \code{input$resourceid} value.
}
\description{
provides a shiny server module for the bfabric system.
It is assumes that the \code{exec/flask_bfabric_sample.py} programm is running.

\code{ssh-keygen -f $PWD/bfabricShiny.key -t rsa} will generate
the key files.
}
\details{
t.b.d.
\enumerate{
\item add \code{bf <- callModule(bfabric, "bfabric8", applicationid = c(155))}
\item follow the instructions \code{\link{bfabricInput}}
}
}
\references{
\url{https://doi.org/10.1145/1739041.1739135}
}
\seealso{
\itemize{
\item \url{http://fgcz-bfabric.uzh.ch}
\item \url{http://fgcz-svn.uzh.ch/repos/scripts/trunk/linux/bfabric/apps/python}
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2017
}
