% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfabric.R
\name{bfabric}
\alias{bfabric}
\title{Shiny server module for the bfabric using bfabricPy}
\usage{
bfabric(
  input,
  output,
  session,
  applicationid = NULL,
  resoucepattern = ".*",
  resourcemultiple = FALSE,
  token = NULL
)
}
\arguments{
\item{input}{default module input}

\item{output}{default module output}

\item{session}{module session}
}
\value{
check the \code{input$resourceid} value.
}
\description{
provides a shiny server module for the bfabric system.
It is assumes that the \code{bfabric_flask.py} program is running, e.g.,
\code{BFABRICPY_CONFIG_ENV=FGCZSTABLE bfabric_flask.py}
}
\details{
ensure file staging via ssh if not NFS is available
\code{ssh-keygen -f $PWD/bfabricShiny.key -t rsa} will generate
the key files.

\enumerate{

\item add \code{bf <- callModule(bfabric, "bfabric8", applicationid = c(155))}

\item follow the instructions \code{\link{bfabricInput}}

}
}
\references{
\url{https://www.degruyterbrill.com/document/doi/10.1515/jib-2022-0031/html}
}
\seealso{
\itemize{
\item \url{https://github.com/fgcz/bfabricPy}
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2017, 2025
}
