% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{bfabric_upload_file}
\alias{bfabric_upload_file}
\title{Generate a workunit and upload a resource (file) to an internal bfabric
storage.}
\usage{
bfabric_upload_file(
  login = NULL,
  webservicepassword = NULL,
  projectid = 3000,
  applicationid = 217,
  status = "pending",
  description = "",
  inputresourceid = NULL,
  workunitname = "MaxQuant result",
  resourcename = "MaxQuant report",
  file_content = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword}

\item{projectid}{a containerid (project id or order id)}

\item{applicationid}{a application id}

\item{status}{in \code{c('AVAILABLE', 'FAILED', 'PENDING')}
default is 'pending'.}

\item{inputresourceid}{an integer or a list of integer inputresourceIds. 
Of note, this works only for  succeeding/preceding applications.
Default is set to \code{NULL}.}

\item{workunitname}{the workunit name}

\item{resourcename}{the reosurce name}

\item{file_content}{a BLOB containing the content}
}
\value{
the workunit id
}
\description{
Generate a workunit and upload a resource (file) to an internal bfabric
storage.
}
