% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{bfabric_upload_file}
\alias{bfabric_upload_file}
\title{upload resouce to bfabric}
\usage{
bfabric_upload_file(login, webservicepassword, projectid = 1000,
  applicationid = 217, status, inputresource = NULL,
  workunitname = "MaxQuant result", resourcename = "MaxQuant report",
  file_content = NULL)
}
\arguments{
\item{login}{}

\item{webservicepassword}{}

\item{projectid}{}

\item{applicationid}{}

\item{status}{}

\item{workunitname}{}

\item{resourcename}{}

\item{file_content}{}
}
\description{
upload resouce to bfabric
}
