% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfabric.R
\name{bfabricInput}
\alias{bfabricInput}
\title{Shiny UI module}
\usage{
bfabricInput(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module
}
\references{
\url{https://doi.org/10.1145/1739041.1739135}
}
\seealso{
\url{http://fgcz-bfabric.uzh.ch}
}
