% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{createWorkunit}
\alias{createWorkunit}
\title{inputresource' = list(inputresource),}
\usage{
createWorkunit(
  login,
  webservicepassword,
  projectid,
  applicationid,
  inputresource,
  status = "available",
  description = "",
  name
)
}
\description{
inputresource' = list(inputresource),
}
