% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{.createWorkunit}
\alias{.createWorkunit}
\alias{bfabricUploadFile}
\title{Create a workunit}
\usage{
.createWorkunit(
  login = NULL,
  webservicepassword = NULL,
  containerid = NULL,
  applicationid = NULL,
  inputresourceid = NULL,
  status = "PENDING",
  description = "",
  name = ""
)
}
\description{
Create a workunit
}
\examples{
\dontrun{
bfabricShiny:::.createWorkunit(login, webservicepassword, containerid=3000, applicationid=500, inputresourceid=list(1753925, 1753924))
}

}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2021
}
