% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_queue.R
\name{generate_queue}
\alias{generate_queue}
\title{FGCZ mass spec queue generator}
\usage{
generate_queue(x, foldername = "", projectid = 1000, area = "Proteomics",
  instrument = "FUSION_1", username = "cpanse", how.often = 2,
  how.many = 1, nr.methods = 2, nr.replicates = 3,
  showcondition = FALSE, qc.type = 1, hplc = "", method = "default",
  pathprefix = "D:\\\\Data2San", pathprefixsep = "\\\\")
}
\arguments{
\item{x}{}

\item{foldername}{}

\item{projectid}{}

\item{area}{}

\item{instrument}{}

\item{username}{}

\item{how.often}{}

\item{how.many}{}

\item{nr.methods}{}

\item{nr.replicates}{}

\item{showcondition}{}

\item{qc.type}{}

\item{hplc}{}

\item{method}{}

\item{pathprefix}{}

\item{pathprefixsep}{}
}
\value{
a instrument configuration as \code{data.frame}.
}
\description{
FGCZ mass spec queue generator
}
