% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{getApplications}
\alias{getApplications}
\title{Title}
\usage{
getApplications(login, webservicepassword)
}
\arguments{
\item{login}{}

\item{webservicepassword}{}
}
\description{
Title
}
\examples{
\dont{
A <- getApplications(login, webservicepassword)
bfabricApplication <- data.frame(id = sapply(A, function(x){x$`_id`}), name = sapply(A, function(x){x$name}))
bfabricApplication <- bfabricApplication[order(bfabricApplication$id),]
write.table(bfabricApplication, file="./data/application.csv")
}

}
