% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{getResources}
\alias{getResources}
\title{get all resources of a (login, project)}
\usage{
getResources(login, webservicepassword, workunitid)
}
\arguments{
\item{login}{}

\item{webservicepassword}{}

\item{project}{}
}
\value{
a vector of resource ids
}
\description{
get all resources of a (login, project)
}

