% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{read}
\alias{read}
\title{read method to access bfabric REST}
\usage{
read(
  login = NULL,
  webservicepassword = NULL,
  endpoint = "workunit",
  query,
  posturl = "http://localhost:5000/",
  as_data_frame = FALSE
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{query}{e,g, \code{list(containerid = 3000)}}

\item{posturl}{POST url, default is \code{'http://localhost:5000/q'}.}

\item{as_data_frame}{if TRUE it returns a data.frame object.}
}
\value{
a nested list object
}
\description{
read method to access bfabric REST
}
\examples{
# ensure you have login and password
Rprofile <- file.path(Sys.getenv("HOME"), ".Rprofile")
source(Rprofile)

res.sample <- bfabricShiny::read(login = login,
   webservicepassword = webservicepassword,
   endpoint = 'sample',
   query = list(containerid = 27053),
   as_data_frame = TRUE)
   
res.sample |> subset(select=c("_id", "name", "groupingvar.name"))
   
res.file <- bfabricShiny::read(login, webservicepassword,
  endpoint = 'resource',
  query = list('filechecksum' = '127f0c5b6352a326f9a6c8458d59d921'),
  )
  
res.workunit <- bfabricShiny::read(login, webservicepassword, 
    endpoint='workunit',
    query = list('status' = 'pending'),
    as_data_frame = TRUE)
      

 ## query metadata
 Q <- bfabricShiny::read(login, webservicepassword,
   endpoint = 'resource',
   query = list('workunitid' = 163763))
   
}
