% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{readPages}
\alias{readPages}
\title{read function which supports pages}
\usage{
readPages(
  login = NULL,
  webservicepassword = NULL,
  endpoint = "workunit",
  query = list(),
  posturl = "http://localhost:5000/read",
  maxpages = 10
)
}
\arguments{
\item{login}{}

\item{webservicepassword}{}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{query}{e,g, \code{list(containerid = 3000)}}

\item{posturl}{where the flask server is working}

\item{maxpages}{max number of supported pages to}
}
\value{
a list
}
\description{
read function which supports pages
}
\examples{
\dontrun{
  fraction <- readPages(login = login,
    webservicepassword = webservicepassword,
    endpoint = 'sample',
    query = list( attribute = list(name = 'fraction',
        value = 'true'))
        }
}
