% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{save}
\alias{save}
\title{Saves an object in bfabric}
\usage{
save(
  login = NULL,
  webservicepassword = NULL,
  endpoint = "workunit",
  posturl = "http://localhost:5000/",
  query = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{endpoint}{the endpoint, e.g., \code{'sample'}}

\item{posturl}{where the flask server is working}

\item{query}{e,g, \code{list(containerid = 3000)}}
}
\value{
bfabric json object.
}
\description{
Saves an object in bfabric
}
\examples{
\dontrun{
bfabricShiny::save(login, webservicepassword , 'workunit',
  posturl=bfabricposturl,
  list(applicationid=212, description='test2', containerid=3000))
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2016-2023, MdE 2023-03-17
}
