% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{save}
\alias{save}
\title{Saves an object in bfabric}
\usage{
save(login, webservicepassword, endpoint = "workunit", query)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword}

\item{endpoint}{bfabric endpoint}

\item{query}{list object to be saved in bfabric.}
}
\description{
Saves an object in bfabric
}
\examples{
\dontrun{
bfabricShiny::save(login, webservicepassword , 'workunit',
  list(id=254893, description='test2',
  inputresourceid=list(1753925, 1753924))
}
}
