% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz.R
\name{uploadResource}
\alias{uploadResource}
\title{Generate a workunit and upload a resource (file)  to a internal bfabric
storage}
\usage{
uploadResource(
  login = NULL,
  webservicepassword = NULL,
  posturl = NULL,
  containerid = 3000,
  applicationid = 212,
  status = "PENDING",
  description = "",
  inputresourceid = NULL,
  workunitname = "bfabricShiny result",
  resourcename = "bfabricShiny report",
  file = NULL
)
}
\arguments{
\item{login}{bfabric login}

\item{webservicepassword}{bfabric webservicepassword,
visible when you check your user details in the bfabric system.}

\item{posturl}{where the flask server is working}

\item{containerid}{a containerid (project id or order id)}

\item{applicationid}{a application id}

\item{status}{in \code{c('AVAILABLE', 'FAILED', 'PENDING')}
default is 'PENDING'.}

\item{description}{free text, default is is empty.}

\item{inputresourceid}{an integer or a list of integer of inputresourceIds. 
Of note, this works only for  succeeding/preceding applications.
Default is set to \code{NULL}.}

\item{workunitname}{the workunit name}

\item{resourcename}{the reosurce name}

\item{file}{a BLOB containing the content}
}
\value{
returns a nested list containing the workunit and the resource
object.
}
\description{
Generate a workunit and upload a resource (file)  to a internal bfabric
storage
}
\examples{
fRp <- file.path(Sys.getenv("HOME"), ".Rprofile")
source(fRp, local=TRUE)
tf <- tempfile()
write.csv(iris, file=tf)
bfabricShiny::uploadResource(login, webservicepassword, bfabricposturl,
   containerid = 3000,
   status = 'PENDING',
   description = "generated by a test run",
   applicationid = 212,
   workunitname = "bfabricShiny example",
   resourcename = "R's iris data",
   file = tf)
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2016-2023, MdE 2023-03-17
}
